+++ homeassistant/components/mqtt/vacuum/__init__.py
@@ -15,50 +15,18 @@ from homeassistant.components.mqtt.discovery import (
     clear_discovery_hash,
 )
 from homeassistant.helpers.dispatcher import async_dispatcher_connect
+from .schema import CONF_SCHEMA, LEGACY, STATE, MQTT_VACUUM_SCHEMA
+from .schema_legacy import PLATFORM_SCHEMA_LEGACY, async_setup_entity_legacy
+from .schema_state import PLATFORM_SCHEMA_STATE, async_setup_entity_state
 
 _LOGGER = logging.getLogger(__name__)
-
-CONF_SCHEMA = "schema"
-LEGACY = "legacy"
-STATE = "state"
 
 
 def validate_mqtt_vacuum(value):
     """Validate MQTT vacuum schema."""
-    from . import schema_legacy
-    from . import schema_state
-
-    schemas = {
-        LEGACY: schema_legacy.PLATFORM_SCHEMA_LEGACY,
-        STATE: schema_state.PLATFORM_SCHEMA_STATE,
-    }
+    schemas = {LEGACY: PLATFORM_SCHEMA_LEGACY, STATE: PLATFORM_SCHEMA_STATE}
     return schemas[value[CONF_SCHEMA]](value)
 
-
-def services_to_strings(services, service_to_string):
-    """Convert SUPPORT_* service bitmask to list of service strings."""
-    strings = []
-    for service in service_to_string:
-        if service & services:
-            strings.append(service_to_string[service])
-    return strings
-
-
-def strings_to_services(strings, string_to_service):
-    """Convert service strings to SUPPORT_* service bitmask."""
-    services = 0
-    for string in strings:
-        services |= string_to_service[string]
-    return services
-
-
-MQTT_VACUUM_SCHEMA = vol.Schema(
-    {
-        vol.Optional(CONF_SCHEMA, default=LEGACY): vol.All(
-            vol.Lower, vol.Any(LEGACY, STATE)
-        )
-    }
-)
 
 PLATFORM_SCHEMA = vol.All(
     MQTT_VACUUM_SCHEMA.extend({}, extra=vol.ALLOW_EXTRA), validate_mqtt_vacuum
@@ -95,13 +63,7 @@ async def _async_setup_entity(
     config, async_add_entities, config_entry, discovery_hash=None
 ):
     """Set up the MQTT vacuum."""
-    from . import schema_legacy
-    from . import schema_state
-
-    setup_entity = {
-        LEGACY: schema_legacy.async_setup_entity_legacy,
-        STATE: schema_state.async_setup_entity_state,
-    }
+    setup_entity = {LEGACY: async_setup_entity_legacy, STATE: async_setup_entity_state}
     await setup_entity[config[CONF_SCHEMA]](
         config, async_add_entities, config_entry, discovery_hash
     )
