+++ homeassistant/components/mqtt/vacuum/schema_state.py
@@ -46,7 +46,7 @@ from homeassistant.components.mqtt import (
     CONF_QOS,
 )
 
-from . import MQTT_VACUUM_SCHEMA, services_to_strings, strings_to_services
+from .schema import MQTT_VACUUM_SCHEMA, services_to_strings, strings_to_services
 
 _LOGGER = logging.getLogger(__name__)
